WIP on block string support.
First pass at block string support.  Not sufficient

This doesn’t trim leading indentation (or trailing whitespace) from the lines inside the block string, which is expected and the specific requirements are detailed in the graphql spec.

Added the algorithm/notes from the spec and a blockStringValue() empty function to implement this in.

Saving this old work because it works other than that.