fix: Encapsulates subscription steam consumption
This fixes an issue observed in downstream code, where awaiting a GraphQL subscription with `makeAsyncIterator().next()` results in an error of `Fatal error: attempt to await next() on more than one task`. This encapsulates the consumption of the underlying resolver stream in a single cancellable task so that downstream code does not share the underlying stream iterator.