refactor: Removes unnecessary DispatchQueue
This is unnecessary because we only ever write the property in its entirety without appending or modifying it in any way. This change may cause us to validate the schema multiple times concurrently early in our start up process, (which should end up with the same result), but the code simplification seems worth it.